maker-desktop/host/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.1",
    "fastapi>=0.116.1",
    "ffmpeg>=1.4",
    "funasr>=1.2.6",
    "openai>=1.99.1",
    "pydub>=0.25.1",
    "python-multipart>=0.0.20",
    "sounddevice>=0.5.2",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.35.0",
]

[tool.aerich]
tortoise_orm = "config.cfg_tortoise_orm"
location = "./migrations"
src_folder = "./."
